requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-chartit==0.2.9
django-chartjs==2.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-rest-auth==0.9.5
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
frozenlist==1.4.0
idna==3.4
isort==5.12.0
Markdown==3.5
multidict==6.0.4
phonenumbers==8.13.24
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
